refactor(pricing): adopt the reconciled Decimal cost-math core
The cost arithmetic moves to core/metered_pricing.py, adopted from the
platform's reconciled core: Decimal end to end, both cached-token
conventions represented, and the convention a caller speaks is an
argument with no default. The pricing lookup chain is unchanged.

The rate columns on model_pricing and organization_model_pricing become
NUMERIC(18, 8) and usage_logs.cost becomes NUMERIC(18, 6), the scale
mozilla-ai/otari-ai#1751 settled for the accounting truth. A cost is
exact until it becomes a settled total, then rounded once, half-up, to
the micro-dollar.

Fixes #661.

+++ scripts/seed_usage_smoke.py
@@ -122,7 +122,7 @@ for tool in list(TOOL_UNIT_RATES):
         .order_by(ModelPricing.effective_at.desc())
         .first()
     )
-    TOOL_UNIT_RATES[tool] = (row.input_price_per_million / 1_000_000) if row else None
+    TOOL_UNIT_RATES[tool] = float(row.input_price_per_million / 1_000_000) if row else None
 print(f"tool rates in effect: {TOOL_UNIT_RATES}")
 
 now = datetime.now(UTC)
